p2p: batch publish data column sidecars
a relatively small change to optimize network send order.

Without this, network writes tend to prioritize sending data for one column
to all peers before sending data for later columns (e.g for two columns
and 4 peers per column it would send A,A,A,A,B,B,B,B). With batch
publishing we can change the write order to round robin across columns
(e.g. A,B,A,B,A,B,A,B).

In cases where the process is sending at a rate over the network limit,
this approach allows at least some copies of the column to propagate
through the network. In early simulations with bandwidth limits of
50mbps for the publisher, this improved dissemination by ~20-30%.